Update - 290123
- Debris collection works fine now
- Shop shows but doesn't work yet

+++ AlphaEngine/Player.h
@@ -21,7 +21,7 @@ struct Player
 	double			beam_width;
 	Planets			current_planet;
 	int				current_capacity; // probably can shorten
-	double			max_capacity;
+	int				max_capacity;
 	double			texture;
 	double			velocity;
